requirements.txt
bs4
tqdm
poetry
requests_toolbelt
